refactor: move getFeatureTopics into TopicList
